fix hash func bug

+++ sperl_hash_func.c
@@ -8,10 +8,10 @@ int64_t SPerl_HASH_FUNC_calc_hash(SPerl* sperl, const char* str, int64_t len) {
   (void)sperl;
   
   const char* str_tmp = str;
-  int64_t hash = 5381;
+  uint32_t hash = 5381;
   while (len--) {
     hash = ((hash << 5) + hash) + (uint8_t) *str_tmp++;
   }
   
-  return hash;
+  return (int64_t)hash;
 }
